agents: move out of exp
