+++ agents/doc.go
@@ -4,4 +4,4 @@
 // a response corresponding to an “action” to take and a corresponding
 // “action input”. Alternatively the agent can return a finish with the
 // finished answer to the query.
-package agent
+package agents
