+++ agents/executor/doc.go
