+++ agents/executor/errors.go
