+++ tools/serpapi/doc.go
@@ -1,0 +1,3 @@
+// Package serpapi contains an implementation of the tool interface with the
+// serapi.
+package serpapi
